First pass of removing docker functions/processes/etc.
