+++ src/dashboard/app/app.html
@@ -28,12 +28,6 @@
 	<script>
 	 const dockerName = {
 	     "#docker-point-node": "point_node",
-	     /* "#docker-contracts": "contract_deployer",
-		"#docker-pgadmin": "pgadmin",
-		"#docker-database": "database",
-		"#docker-blockchain": "blockchain_node",
-		"#docker-owner": "website_owner",
-		"#docker-visitor": "website_visitor", */
 	 }
 	 let isFirefoxInstalled = false;
 	 let isPointNodeRunning = false;
