+++ src/firefox/index.js
@@ -89,25 +89,29 @@ class Firefox {
 
     async launch() {
 	const osAndArch = helpers.getOSAndArch();
-        const cmd = await this.getBinPath(osAndArch);
+        const bin = await this.getBinPath(osAndArch);
         const profile_path = path.join(await helpers.getHomePath(), ".point/keystore/profile");
         const flags = "--profile "+profile_path;
-	let webext_binary = '';
-	if (osAndArch == 'win32' || osAndArch == 'win64') {
-		webext_binary = 'web-ext';
-	} else {
-		webext_binary = path.join(await helpers.getHomePath(), ".point/src/pointnetwork-dashboard/node_modules/web-ext/bin/web-ext");
-	}
-        // const webext_binary = path.join(await helpers.getHomePath(), ".point/src/pointnetwork-dashboard/node_modules/web-ext/bin/web-ext");
-        const ext_path = path.join(await helpers.getHomePath(), ".point/src/pointsdk/dist/prod"); // should contain manifest.json
-	let webext = '';
-	if (osAndArch == 'win32' || osAndArch == 'win64') {
-		webext = `"${webext_binary}" run "--firefox=${cmd}" "--firefox-profile=${profile_path}" --keep-profile-changes "--source-dir=${ext_path}" --url https://point`;
-	} else {
-		webext = `${webext_binary} run --firefox="${cmd}" --firefox-profile ${profile_path} --keep-profile-changes --source-dir ${ext_path} --url https://point`;
-	}
 
-        exec(webext, (error, stdout, stderr) => {
+	// let webext_binary = '';
+	// if (osAndArch == 'win32' || osAndArch == 'win64') {
+	//     webext_binary = 'web-ext';
+	// } else {
+	//     webext_binary = path.join(await helpers.getHomePath(), ".point/src/pointnetwork-dashboard/node_modules/web-ext/bin/web-ext");
+	// }
+
+        const cmd = `${bin} --profile ${profile_path} --keep-profile-changes --url https://point`;
+              
+        
+        // const ext_path = path.join(await helpers.getHomePath(), ".point/src/pointsdk/dist/prod"); // should contain manifest.json
+	// let webext = '';
+	// if (osAndArch == 'win32' || osAndArch == 'win64') {
+	// 	webext = `"${webext_binary}" run "--firefox=${cmd}" "--firefox-profile=${profile_path}" --keep-profile-changes "--source-dir=${ext_path}" --url https://point`;
+	// } else {
+	// 	webext = `${webext_binary} run --firefox="${cmd}" --firefox-profile ${profile_path} --keep-profile-changes --source-dir ${ext_path} --url https://point`;
+	// }
+
+        exec(cmd, (error, stdout, stderr) => {
             // win.webContents.send("firefox-closed");
             if (error) {
                 console.log(`error: ${error.message}`);
@@ -253,6 +257,7 @@ pref('network.proxy.autoconfig_url', '${pacFile}');
 pref('security.enterprise_roots.enabled', true);
 pref('network.captive-portal-service.enabled', false);
 pref('browser.tabs.drawInTitlebar', true);
+pref("extensions.autoDisableScopes", 0);
 `;
         const prefPath = await this.getPrefPath(osAndArch);
         const appPath = await this.getAppPath(osAndArch);
