+++ src/installer/app/app.html
@@ -30,8 +30,6 @@
                 <p>Welcome to Point Network installer. Please review the components to be installed and click "Start".</p>
 
                 <ul>
-                    <li>WSL2</li>
-                    <li>Docker & docker-compose</li>
                     <li>Point Network</li>
                 </ul>
