+++ src/installer/service/index.js
@@ -2,7 +2,7 @@ import * as fsExtra from "fs-extra";
 import * as fs from "fs";
 import * as path from "path";
 import helpers, {getOSAndArch} from "../../helpers";
-import docker from "../../docker";
+import pointnode from "../../pointnode";
 import Welcome from "../../welcome";
 import Dashboard from "../../dashboard";
 const util = require('util');
@@ -33,9 +33,7 @@ class InstallerService {
         Object.assign(this.steps, {
             'Create ~/.point directory': this.makePointDirectory,
             'Create ~/.point/src directory': this.makePointSrcDirectory,
-            'Clone PointNetwork repositories': this.cloneRepos,
-            'Install Docker': this.installDocker,
-            'Running Docker Containers': this.runDockerCompose,
+            'Clone PointNetwork repositories': this.cloneRepos
         });
     }
 
@@ -92,54 +90,6 @@ class InstallerService {
             return false;
         }
         sudo.exec('wsl --install');
-    }
-
-    async installDocker(testRun = false) {
-        if (testRun) {
-            return docker.isInstalled();
-        }
-        docker.install(this.osAndArch);
-
-        // if (this.isWindows64()) {
-        //     const url = "https://desktop.docker.com/win/main/amd64/Docker%20Desktop%20Installer.exe";
-
-        // }
-
-        // const response = await axios.get('https://github.com/docker/compose/releases/latest', {
-        //     headers: { accept: 'application/json' },
-        //     data: {}
-        // });
-        // /**
-        //  *     data: {
-        //           id: 42700380,
-        //           tag_name: '1.29.2',
-        //           update_url: '/docker/compose/releases/tag/1.29.2',
-        //           update_authenticity_token: 'Z0iLO6fLimW0WB3tuZhyBg1j82eJg+6mYCtBAwY7mFs4Wv4BRsgSByzSMj4yBzk3a/6Vzicvrzh1AefJAiGEFQ==',
-        //           delete_url: '/docker/compose/releases/tag/1.29.2',
-        //           delete_authenticity_token: 'SQoTXN7DrXP2oZIMXQgpJ3ilrg2OvRJZPv4a60wXcjuAk98ceaZ44IFEKvaY6JYHy73Ty8wQCmk2cWwAjVGGfA==',
-        //           edit_url: '/docker/compose/releases/edit/1.29.2'
-        //         }
-        //  */
-        // const latest_tag = response.data.tag_name;
-        // this._log('Latest docker/compose release: '+latest_tag); // todo: escape
-
-        // sudo.exec('curl -L '+this._quote("https://github.com/docker/compose/releases/download/"+latest_tag+"/docker-compose-$(uname -s)-$(uname -m)")+' -o /usr/local/bin/docker-compose');
-    }
-
-    // TODO: Create `helpers.getDockerComposePath` and use that.
-    // TODO: Move `this._exec` to `helpers` or something.
-    // TODO: Just noticed `../welcome` has exact copies of _exec. Refactor before doing above TODOs.
-    async runDockerCompose(testRun = false) {
-        // TODO: `this._execAndGetOutput` doesn't return output, so it can't be used to test.
-        // if (testRun) return await this._execAndGetOutput('docker-compose ps -q');
-        if (testRun) return false;
-        const pnPath = await helpers.getPNPath(this.osAndArch);
-        const composePath = helpers.fixPath(this.osAndArch, path.join(pnPath, 'docker-compose.yaml'));
-        // TODO: `docker-compose.dev.yaml` is buggy at the moment.
-        // const composeDevPath = helpers.fixPath(osAndArch, path.join(pnPath, 'docker-compose.dev.yaml'));
-        // const cmd = `docker-compose -f ${composePath} -f ${composeDevPath} up -d`;
-        const cmd = `docker-compose -f ${composePath} up -d`;
-        await this._exec(cmd);
     }
 
     async cloneRepos(testRun = false) {
@@ -259,20 +209,6 @@ class InstallerService {
         const homedir = await helpers.getHomePath();
         return path.join(homedir, ...paths);
     }
-
-    // async getHealthCmd(osAndArch, containerName) {
-    //     const pnPath = await helpers.getPNPath(helpers.getOSAndArch());
-    //     const composePath = helpers.fixPath(osAndArch, path.join(pnPath, 'docker-compose.yaml'));
-    //     const composeDevPath = helpers.fixPath(osAndArch, path.join(pnPath, 'docker-compose.dev.yaml'));
-    //
-    //     const cmd = `docker inspect --format "{{json .State.Health}}" $(docker-compose -f ${composePath} -f ${composeDevPath} ps -q ${containerName})`;
-    //     if (osAndArch == 'win32' || osAndArch == 'win64') {
-    //         return `wsl ${cmd}`;
-    //     }
-    //     return cmd;
-    // }
-
-
 }
 
 export default InstallerService;
